requirements.txt
pydantic==2.12.4
python-dotenv==1.2.1
requests==2.32.5